cli/requirements.txt
git+https://github.com/insight-stream/insight_stream.git
python-dotenv